% Generated by roxygen2 (4.0.1): do not edit by hand
\name{partition}
\alias{partition}
\title{Partitions an input vector a around a pivot at position pindex}
\usage{
partition(a, pindex)
}
\arguments{
\item{a}{A positive integer numeric vector}

\item{pindex}{The position of the pivot}
}
\value{
A list with the input array where all elements less than or equal to
the pivot element will be located before the pivot, together with the new
position for the pivot element
}
\description{
Partitions an input vector a around a pivot at position pindex
}
\details{
TODO
}
\author{
Jose A. Dianes
}

