% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pivot_first}
\alias{pivot_first}
\title{Selects the first element as a pivot}
\usage{
pivot_first(a)
}
\arguments{
\item{a}{A positive integer numeric vector}
}
\value{
The index 1
}
\description{
Selects the first element as a pivot
}
\details{
TODO
}
\author{
Jose A. Dianes
}

