% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pivot_last}
\alias{pivot_last}
\title{Selects the last element as a pivot}
\usage{
pivot_last(a)
}
\arguments{
\item{a}{A positive integer numeric vector}
}
\value{
The index lengt(a)
}
\description{
Selects the last element as a pivot
}
\details{
TODO
}
\author{
Jose A. Dianes
}

