% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pivot_median_3}
\alias{pivot_median_3}
\title{Selects the median of the first, last, and middle elements as a pivot}
\usage{
pivot_median_3(a)
}
\arguments{
\item{a}{A positive integer numeric vector}
}
\value{
The median of the first, last, and middle elements as a pivot
}
\description{
Selects the median of the first, last, and middle elements as a pivot
}
\details{
TODO
}
\author{
Jose A. Dianes
}

