% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pivot_random}
\alias{pivot_random}
\title{Selects a random element as a pivot, using sample()}
\usage{
pivot_random(a)
}
\arguments{
\item{a}{A positive integer numeric vector}
}
\value{
A random index using sample()
}
\description{
Selects a random element as a pivot, using sample()
}
\details{
TODO
}
\author{
Jose A. Dianes
}

