% Generated by roxygen2 (4.0.1): do not edit by hand
\name{quicksort_counter}
\alias{quicksort_counter}
\title{Sorts the input numeric vector using QUICKSORT and counts the number of comparisons
Intended to be used to compare different strategies of pivot selection}
\usage{
quicksort_counter(a, PIVOTFUN = pivot_median_3)
}
\arguments{
\item{a}{An integer numeric vector}

\item{PIVOTFUN}{The function that selects the pivot element index}
}
\value{
A list with the input vector already sorted and a count of the
number of comparisons
}
\description{
Sorts the input numeric vector using QUICKSORT and counts the number of comparisons
Intended to be used to compare different strategies of pivot selection
}
\details{
TODO
}
\author{
Jose A. Dianes
}

