% Generated by roxygen2 (4.0.1): do not edit by hand
\name{quicksort_mc}
\alias{quicksort_mc}
\title{Sorts the input numeric vector using QUICKSORT and foreach parallel}
\usage{
quicksort_mc(a, PIVOTFUN = pivot_median_3, ncores = 2)
}
\arguments{
\item{a}{An integer numeric vector}

\item{PIVOTFUN}{The function that selects the pivot element index}
}
\value{
A list with the input vector already sorted
}
\description{
Sorts the input numeric vector using QUICKSORT and foreach parallel
}
\details{
TODO
}
\author{
Jose A. Dianes
}

